+++ client/src/MainPage/MainPage.tsx
@@ -7,12 +7,16 @@ import Header from "../components/Header";
 import UserInfo from "../components/Header/UserInfo";
 import FullScreenModal from "../components/modal/FullScreenModal";
 import Footer from "../components/Footer";
+import { Toast } from "../components/Toast/Toast";
+import TOAST from "../components/Toast/ToastList";
 
+import useError from "../hooks/useError";
 import userStore from "../store/user.store";
+import toastStore from "../store/toast.store";
+import mainStore from "../store/main.store";
 import FloatLayout from "../layouts/FloatLayout";
 import Splash from "./components/SplashScreen/SplashScreen";
 import ThemeSeletor from "../components/ThemeSelector";
-import { Toast } from "../components/Toast/Toast";
 
 function CreateMonumentButton({ showModal }: { showModal: () => void }) {
   const isLoggedIn = userStore((store) => store.isLoggedIn);
@@ -26,6 +30,13 @@ function CreateMonumentButton({ showModal }: { showModal: () => void }) {
 
 export default function MainPage() {
   const [show, setShow] = useState<boolean>(false);
+  const addToast = toastStore((store) => store.addToast);
+  const setShowSplash = mainStore((store) => store.setShowSplash);
+
+  useError((reason) => {
+    addToast(TOAST.ERROR(reason));
+    setShowSplash(false);
+  });
 
   function showModal() {
     setShow(true);
